+++ docs/pages/release_notes.md
@@ -31,7 +31,12 @@ This is a {{ site.pmd.release_type }} release.
 * {% jdoc java::lang.java.JavaLanguageHandler %}
 * {% jdoc java::lang.java.JavaLanguageParser %}
 * {% jdoc java::lang.java.JavaDataFlowHandler %}
-
+* Implementations of {% jdoc core::lang.rule.RuleViolationFactory %} in each
+  language module, eg {% jdoc java::lang.java.rule.JavaRuleViolationFactory %}.
+  See javadoc of {% jdoc core::lang.rule.RuleViolationFactory %}.
+* Implementations of {% jdoc core::RuleViolation %} in each language module,
+  eg {% jdoc java::lang.java.rule.JavaRuleViolation %}. See javadoc of
+  {% jdoc core::RuleViolation %}.
 
 ##### For removal